tests: Add test for create SRD QP
Add test create SRDQP for class SRDQP which reperesnt EFA's direct verb
efadv_create_driver_qp.
